refactor: diaryDTO
