+++ watomate/Watomate/Sources/service/NetworkResult.swift
